requirements.txt
google-generativeai>=0.3.1
langchain-google-genai>=0.0.7
streamlit>=1.24.0
python-dotenv>=1.0.0
langchain>=0.1.0